[web] refactor text to own module
