+++ piet-web/src/lib.rs
@@ -1,5 +1,5 @@
 //! The Web Canvas backend for the Piet 2D graphics abstraction.
-mod grapheme;
+mod text;
 
 use std::borrow::Cow;
 use std::fmt;
@@ -15,13 +15,12 @@ use web_sys::{
 use piet::kurbo::{Affine, PathEl, Point, Rect, Shape};
 
 use piet::{
-    Color, Error, FixedGradient, Font, FontBuilder, GradientStop, HitTestMetrics, HitTestPoint,
-    HitTestTextPosition, ImageFormat, InterpolationMode, IntoBrush, LineCap, LineJoin,
-    RenderContext, StrokeStyle, Text, TextLayout, TextLayoutBuilder,
+    Color, Error, FixedGradient, GradientStop,
+    ImageFormat, InterpolationMode, IntoBrush, LineCap, LineJoin,
+    RenderContext, StrokeStyle,
 };
-use unicode_segmentation::UnicodeSegmentation;
 
-use crate::grapheme::point_x_in_grapheme;
+pub use text::{WebTextLayout};
 
 pub struct WebRenderContext<'a> {
     ctx: &'a mut CanvasRenderingContext2d,
@@ -46,43 +45,12 @@ pub enum Brush {
     Gradient(CanvasGradient),
 }
 
-#[derive(Clone)]
-pub struct WebFont {
-    family: String,
-    weight: u32,
-    style: FontStyle,
-    size: f64,
-}
-
-pub struct WebFontBuilder(WebFont);
-
-pub struct WebTextLayout {
-    ctx: CanvasRenderingContext2d,
-    font: WebFont,
-    text: String,
-}
-
-pub struct WebTextLayoutBuilder {
-    ctx: CanvasRenderingContext2d,
-    font: WebFont,
-    text: String,
-}
-
 pub struct WebImage {
     /// We use a canvas element for now, but could be ImageData or ImageBitmap,
     /// so consider an enum.
     inner: HtmlCanvasElement,
     width: u32,
     height: u32,
-}
-
-/// https://developer.mozilla.org/en-US/docs/Web/CSS/font-style
-#[allow(dead_code)] // TODO: Remove
-#[derive(Clone)]
-enum FontStyle {
-    Normal,
-    Italic,
-    Oblique(Option<f64>),
 }
 
 #[derive(Debug)]
@@ -390,33 +358,6 @@ fn set_gradient_stops(dst: &mut CanvasGradient, src: &[GradientStop]) {
     }
 }
 
-impl<'a> Text for WebRenderContext<'a> {
-    type Font = WebFont;
-    type FontBuilder = WebFontBuilder;
-    type TextLayout = WebTextLayout;
-    type TextLayoutBuilder = WebTextLayoutBuilder;
-
-    fn new_font_by_name(&mut self, name: &str, size: f64) -> Self::FontBuilder {
-        let font = WebFont {
-            family: name.to_owned(),
-            size,
-            weight: 400,
-            style: FontStyle::Normal,
-        };
-        WebFontBuilder(font)
-    }
-
-    fn new_text_layout(&mut self, font: &Self::Font, text: &str) -> Self::TextLayoutBuilder {
-        WebTextLayoutBuilder {
-            // TODO: it's very likely possible to do this without cloning ctx, but
-            // I couldn't figure out the lifetime errors from a `&'a` reference.
-            ctx: self.ctx.clone(),
-            font: font.clone(),
-            text: text.to_owned(),
-        }
-    }
-}
-
 impl<'a> WebRenderContext<'a> {
     /// Set the source pattern to the brush.
     ///
@@ -505,177 +446,3 @@ fn byte_to_frac(byte: u32) -> f64 {
     ((byte & 255) as f64) * (1.0 / 255.0)
 }
 
-impl FontBuilder for WebFontBuilder {
-    type Out = WebFont;
-
-    fn build(self) -> Result<Self::Out, Error> {
-        Ok(self.0)
-    }
-}
-
-impl Font for WebFont {}
-
-impl WebFont {
-    fn get_font_string(&self) -> String {
-        let style_str = match self.style {
-            FontStyle::Normal => Cow::from("normal"),
-            FontStyle::Italic => Cow::from("italic"),
-            FontStyle::Oblique(None) => Cow::from("italic"),
-            FontStyle::Oblique(Some(angle)) => Cow::from(format!("oblique {}deg", angle)),
-        };
-        format!(
-            "{} {} {}px \"{}\"",
-            style_str, self.weight, self.size, self.family
-        )
-    }
-}
-
-impl TextLayoutBuilder for WebTextLayoutBuilder {
-    type Out = WebTextLayout;
-
-    fn build(self) -> Result<Self::Out, Error> {
-        self.ctx.set_font(&self.font.get_font_string());
-        Ok(WebTextLayout {
-            ctx: self.ctx,
-            font: self.font,
-            text: self.text,
-        })
-    }
-}
-
-impl TextLayout for WebTextLayout {
-    fn width(&self) -> f64 {
-        //cairo:
-        //self.font.text_extents(&self.text).x_advance
-        self.ctx
-            .measure_text(&self.text)
-            .map(|m| m.width())
-            .expect("Text measurement failed")
-    }
-
-    // first assume one line.
-    // TODO do with lines
-    fn hit_test_point(&self, point: Point) -> HitTestPoint {
-        // internal logic is using grapheme clusters, but return the text position associated
-        // with the border of the grapheme cluster.
-
-        // null case
-        if self.text.len() == 0 {
-            return HitTestPoint::default();
-        }
-
-        // get bounds
-        // TODO handle if string is not null yet count is 0?
-        let end = UnicodeSegmentation::graphemes(self.text.as_str(), true).count() - 1;
-        let end_bounds = match self.get_grapheme_boundaries(end) {
-            Some(bounds) => bounds,
-            None => return HitTestPoint::default(),
-        };
-
-        let start = 0;
-        let start_bounds = match self.get_grapheme_boundaries(start) {
-            Some(bounds) => bounds,
-            None => return HitTestPoint::default(),
-        };
-
-        // first test beyond ends
-        if point.x > end_bounds.trailing {
-            let mut res = HitTestPoint::default();
-            res.metrics.text_position = self.text.len();
-            return res;
-        }
-        if point.x <= start_bounds.leading {
-            return HitTestPoint::default();
-        }
-
-        // then test the beginning and end (common cases)
-        if let Some(hit) = point_x_in_grapheme(point.x, &start_bounds) {
-            return hit;
-        }
-        if let Some(hit) = point_x_in_grapheme(point.x, &end_bounds) {
-            return hit;
-        }
-
-        // Now that we know it's not beginning or end, begin binary search.
-        // Iterative style
-        let mut left = start;
-        let mut right = end;
-        loop {
-            // pick halfway point
-            let middle = left + ((right - left) / 2);
-
-            let grapheme_bounds = match self.get_grapheme_boundaries(middle) {
-                Some(bounds) => bounds,
-                None => return HitTestPoint::default(),
-            };
-
-            if let Some(hit) = point_x_in_grapheme(point.x, &grapheme_bounds) {
-                return hit;
-            }
-
-            // since it's not a hit, check if closer to start or finish
-            // and move the appropriate search boundary
-            if point.x < grapheme_bounds.leading {
-                right = middle;
-            } else if point.x > grapheme_bounds.trailing {
-                left = middle + 1;
-            } else {
-                unreachable!("hit_test_point conditional is exhaustive");
-            }
-        }
-    }
-
-    fn hit_test_text_position(&self, text_position: usize) -> Option<HitTestTextPosition> {
-        // Using substrings, but now with unicode grapheme awareness
-
-        let text_len = self.text.len();
-
-        if text_position == 0 {
-            return Some(HitTestTextPosition::default());
-        }
-
-        if text_position as usize >= text_len {
-            let x = self.width();
-
-            return Some(HitTestTextPosition {
-                point: Point { x, y: 0.0 },
-                metrics: HitTestMetrics {
-                    text_position: text_len,
-                },
-            });
-        }
-
-        // Already checked that text_position > 0 and text_position < count.
-        // If text position is not at a grapheme boundary, use the text position of current
-        // grapheme cluster. But return the original text position
-        // Use the indices (byte offset, which for our purposes = utf8 code units).
-        let grapheme_indices = UnicodeSegmentation::grapheme_indices(self.text.as_str(), true)
-            .take_while(|(byte_idx, _s)| text_position >= *byte_idx);
-
-        if let Some((byte_idx, _s)) = grapheme_indices.last() {
-            let x = self
-                .ctx
-                .measure_text(&self.text[0..byte_idx])
-                .map(|m| m.width())
-                .expect("Text measurement failed");
-
-            Some(HitTestTextPosition {
-                point: Point { x, y: 0.0 },
-                metrics: HitTestMetrics {
-                    text_position: text_position,
-                },
-            })
-        } else {
-            // iterated to end boundary
-            Some(HitTestTextPosition {
-                point: Point {
-                    x: self.width(),
-                    y: 0.0,
-                },
-                metrics: HitTestMetrics {
-                    text_position: text_len,
-                },
-            })
-        }
-    }
-}
